tcpip.c: Run slow poll loop even if no timeouts are present
When non-TCP threads use TCP directly (LWIP_TCPIP_CORE_LOCKING=1) there
might be a case when the TCP thread goes to sleep waiting for a new
message for ever that new timeouts are entered by a second thread but
are never processed because the tcp thread never wakes up.
To prevent this the TCP thread always wakes up from waiting for the next
message after 500ms, and checks again for timeouts.

If a timeout was pending the loop started an active waiting loop without
ever releasing the TCP lock. Return to the top of the function only when
the sleeptime was 0, i.e. a timer was processed immediately. If not
proceed to the bottom of the function where it waits for the pending
next timeout time.